Make initial ao sample direction random
When changing the point of view (moving the camera), the frame counter on which the random seed for ao was reset.
This works great for resetting the sample accumulation buffer, but it force the initial seed for each frame to be the exact same, until the camera stops moving.
I found that annoying and easy to fix, so I gave it a go and simply added a global frame counter to ao noise seed, so the pattern keeps changing even when the accumulation is reset.

TODO : aoControl.frame value should wrap around aoControl.max_samples, to prevent numeric imprecisions with large number. Might not be an actual issue since an int is used rather than a float as a noise seed, but it should be double-checked.